# change logic for setting MG param coarse_grid_solution_type