@kostrzewa 2022-06-03
@marcuspetschlies if you remember when you were testing analysis workloads with the `quda_work_add_actions` branch you were finding crazy slow behaviour of the MG compared to the `quda_work` branch. We suspected an issue in QUDA (and there were also a couple of issues there), but THIS was the actual and main reason...