@Finkenrath 2022-06-03
Ok, that looks good. I will try to check it (need to re-compile QUDA for that).
The heatbath problem seems that this indicates that the rho-shifts are not linear. However in case of DDalphaAMG one looses also performance/increase of iteration count if the shifts are too large, but much later rho ~ 0.1.