@kostrzewa 2022-06-05
See https://github.com/lattice/quda/issues/1287 for updates on the `updateMultigridQuda` regression.