@kostrzewa 2022-06-05
I am unable to reproduce this on my dev machine with dual GTX-1060 (and hence without inter-node comms).